+++ Assets/GameComponent/Scripts/PlayerScript/AttackForPlayer.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class AttackForPlayer : MonoBehaviour
 {
-    public float attackDamage = 20.0f;
+    public int attackDamage = 20;
     public  Vector2 Knockback = new Vector2 (2,1);
 
     // Start is called before the first frame update
